+++ pymob/solvers/base.py
@@ -1,4 +1,5 @@
 import numpy as np
+import xarray as xr
 from typing import Callable, Dict, List, Optional, Sequence, Literal, Tuple
 from frozendict import frozendict
 from dataclasses import dataclass, field
@@ -83,3 +84,61 @@ def mappar(func, parameters, exclude=[], to:Literal["tuple","dict"]="tuple"):
 
     return model_args
 
+
+def create_interpolation(
+        x_in: xr.Dataset, 
+        x: str="time", 
+        factor: float=1e-4, 
+        interpolation: Literal["fill-forward", "linear"] = "fill-forward",
+    ) -> xr.Dataset:
+    """Make the interpolation safe by adding a coordinate just before each 
+    x-value (except the first vaue). The distance between the new and the next
+    point are calculated as a fraction of the previous distance between
+    neighboring points. The corresponding y-values are first set to NaN and then
+    interpolated based on the interpolation method.
+
+    Parameters
+    ----------
+    x_in : xr.Dataset
+        The input dataset which contains a coordinate (x) and a data variable
+        (y)
+    x : str, optional
+        The name of the x coordinate, by default "time"
+    factor : float, optional
+        The distance between the newly added points and the following existing
+        points on the x-scale, by default 1e-4
+    interpolation : Literal["fill-forward", "linear"], optional
+        The interpolation method. In addition to 'fill-forward' and 'linear',
+        any method give in `xr.interpolate_na` can be chosen, by default
+        "fill-forward"
+
+    Returns
+    -------
+    xr.Dataset
+        The interpolated dataset
+    """
+    xs = x_in.coords[x]
+
+    # calculate x values that are located just a little bit smaller than the xs
+    # where "just a little bit" is defined by the distance to the previous x
+    # and a factor. This way the scale of the observations should not matter
+    # and even very differently sized x-steps should be interpolated correctly
+    fraction_before_xs = (
+        xs.isel({x:range(1, len(xs))}).values
+        - xs.diff(dim=x) * factor
+    )
+
+    # create a sorted time vector
+    xs = sorted([*fraction_before_xs.values, *xs.values])
+
+    # add new time indices with NaN values 
+    x_in_reindexed = x_in.reindex({x:xs})
+
+    if interpolation == "fill-forward":
+        # then fill nan values with the previous value (forward-fill)
+        x_in_interpolated = x_in_reindexed.ffill(dim=x, limit=1)
+
+    else:
+        x_in_interpolated = x_in_reindexed.interpolate_na(dim=x, method="linear")
+
+    return x_in_interpolated
